requirements.txt
torch>=1.10.0
torchvision>=0.11.0
transformers>=4.20.0
pillow>=9.0.0
matplotlib>=3.5.0
numpy>=1.20.0
scikit-learn>=1.0.0
pandas>=1.3.0
tqdm>=4.62.0
requests>=2.26.0
jsonschema>=4.4.0
ipykernel>=6.6.0  # For Jupyter notebooks
